Run a simple Servlet
A simple java Servlet.

Important thing to note is <scope>provided</scope> in javax dependency in
pom.xml. Not adding the scope would lead to inclusion of javax jar in the
generated war. This leads to MudServlet is not a Servlet exception.

The path to the servlet was specified in the web.xml file.

Signed-off-by: Venkatesh Shukla <venkatesh.shukla.eee11@iitbhu.ac.in>